requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-WTF==1.2.2
Flask-Bootstrap==3.3.7.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
Pillow==11.1.0
gunicorn==23.0.0
email-validator==2.2.0